Split out semantic models for MSBuild and XML.
