+++ MSBuildProjectTools.sln
@@ -25,6 +25,12 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LanguageServer.SemanticModel.MSBuild", "src\LanguageServer.SemanticModel.MSBuild\LanguageServer.SemanticModel.MSBuild.csproj", "{15BE4E0D-3EBC-40ED-94EA-212A96A81DA3}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LanguageServer.SemanticModel.Xml", "src\LanguageServer.SemanticModel.Xml\LanguageServer.SemanticModel.Xml.csproj", "{277F8A55-417B-4AD0-8AAD-D075886D68BC}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LanguageServer.Common", "src\LanguageServer.Common\LanguageServer.Common.csproj", "{5AD7AEFF-4E5C-474C-8BD6-DEBD354A803C}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -71,6 +77,42 @@ Global
 		{029E8D10-FEDC-4664-ABA9-0A09E9FE89AD}.Release|x64.Build.0 = Release|Any CPU
 		{029E8D10-FEDC-4664-ABA9-0A09E9FE89AD}.Release|x86.ActiveCfg = Release|Any CPU
 		{029E8D10-FEDC-4664-ABA9-0A09E9FE89AD}.Release|x86.Build.0 = Release|Any CPU
+		{15BE4E0D-3EBC-40ED-94EA-212A96A81DA3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{15BE4E0D-3EBC-40ED-94EA-212A96A81DA3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{15BE4E0D-3EBC-40ED-94EA-212A96A81DA3}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{15BE4E0D-3EBC-40ED-94EA-212A96A81DA3}.Debug|x64.Build.0 = Debug|Any CPU
+		{15BE4E0D-3EBC-40ED-94EA-212A96A81DA3}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{15BE4E0D-3EBC-40ED-94EA-212A96A81DA3}.Debug|x86.Build.0 = Debug|Any CPU
+		{15BE4E0D-3EBC-40ED-94EA-212A96A81DA3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{15BE4E0D-3EBC-40ED-94EA-212A96A81DA3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{15BE4E0D-3EBC-40ED-94EA-212A96A81DA3}.Release|x64.ActiveCfg = Release|Any CPU
+		{15BE4E0D-3EBC-40ED-94EA-212A96A81DA3}.Release|x64.Build.0 = Release|Any CPU
+		{15BE4E0D-3EBC-40ED-94EA-212A96A81DA3}.Release|x86.ActiveCfg = Release|Any CPU
+		{15BE4E0D-3EBC-40ED-94EA-212A96A81DA3}.Release|x86.Build.0 = Release|Any CPU
+		{277F8A55-417B-4AD0-8AAD-D075886D68BC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{277F8A55-417B-4AD0-8AAD-D075886D68BC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{277F8A55-417B-4AD0-8AAD-D075886D68BC}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{277F8A55-417B-4AD0-8AAD-D075886D68BC}.Debug|x64.Build.0 = Debug|Any CPU
+		{277F8A55-417B-4AD0-8AAD-D075886D68BC}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{277F8A55-417B-4AD0-8AAD-D075886D68BC}.Debug|x86.Build.0 = Debug|Any CPU
+		{277F8A55-417B-4AD0-8AAD-D075886D68BC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{277F8A55-417B-4AD0-8AAD-D075886D68BC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{277F8A55-417B-4AD0-8AAD-D075886D68BC}.Release|x64.ActiveCfg = Release|Any CPU
+		{277F8A55-417B-4AD0-8AAD-D075886D68BC}.Release|x64.Build.0 = Release|Any CPU
+		{277F8A55-417B-4AD0-8AAD-D075886D68BC}.Release|x86.ActiveCfg = Release|Any CPU
+		{277F8A55-417B-4AD0-8AAD-D075886D68BC}.Release|x86.Build.0 = Release|Any CPU
+		{5AD7AEFF-4E5C-474C-8BD6-DEBD354A803C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5AD7AEFF-4E5C-474C-8BD6-DEBD354A803C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5AD7AEFF-4E5C-474C-8BD6-DEBD354A803C}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{5AD7AEFF-4E5C-474C-8BD6-DEBD354A803C}.Debug|x64.Build.0 = Debug|Any CPU
+		{5AD7AEFF-4E5C-474C-8BD6-DEBD354A803C}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{5AD7AEFF-4E5C-474C-8BD6-DEBD354A803C}.Debug|x86.Build.0 = Debug|Any CPU
+		{5AD7AEFF-4E5C-474C-8BD6-DEBD354A803C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5AD7AEFF-4E5C-474C-8BD6-DEBD354A803C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5AD7AEFF-4E5C-474C-8BD6-DEBD354A803C}.Release|x64.ActiveCfg = Release|Any CPU
+		{5AD7AEFF-4E5C-474C-8BD6-DEBD354A803C}.Release|x64.Build.0 = Release|Any CPU
+		{5AD7AEFF-4E5C-474C-8BD6-DEBD354A803C}.Release|x86.ActiveCfg = Release|Any CPU
+		{5AD7AEFF-4E5C-474C-8BD6-DEBD354A803C}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -79,6 +121,9 @@ Global
 		{04940A47-F57A-47CC-88CE-3B36EDFFAD3E} = {1F2A0703-FB4A-42A9-8CD3-E2F64DF29C6C}
 		{2DF51EC6-ED83-4AEC-9C39-C22510DEBF02} = {1F2A0703-FB4A-42A9-8CD3-E2F64DF29C6C}
 		{029E8D10-FEDC-4664-ABA9-0A09E9FE89AD} = {198F2A53-5FD7-4F6E-A718-DDD95EA1F192}
+		{15BE4E0D-3EBC-40ED-94EA-212A96A81DA3} = {1F2A0703-FB4A-42A9-8CD3-E2F64DF29C6C}
+		{277F8A55-417B-4AD0-8AAD-D075886D68BC} = {1F2A0703-FB4A-42A9-8CD3-E2F64DF29C6C}
+		{5AD7AEFF-4E5C-474C-8BD6-DEBD354A803C} = {1F2A0703-FB4A-42A9-8CD3-E2F64DF29C6C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {BCC91968-D6AB-4FD1-808D-980EE91FAEE2}
