+++ src/LanguageServer.Common/Range.cs
