+++ src/LanguageServer.Common/Utilities/MSBuildHelper.cs
@@ -7,8 +7,6 @@ using System.IO;
 
 namespace MSBuildProjectTools.LanguageServer.Utilities
 {
-    using Utilities;
-
     /// <summary>
     ///     Helper methods for working with MSBuild projects.
     /// </summary>
@@ -153,75 +151,49 @@ namespace MSBuildProjectTools.LanguageServer.Utilities
             );
             if (endPosition == Position.Zero)
                 endPosition = startPosition;
-            
+
             return new Range(startPosition, endPosition);
         }
 
         /// <summary>
-        ///     Get the condition (if any) declared on the element or one of its ancestors.
+        ///     The names of well-known MSBuild properties.
         /// </summary>
-        /// <param name="projectElement">
-        ///     The element.
-        /// </param>
-        /// <returns>
-        ///     The condition, or an empty string if no condition is present on the element or one of its ancestors.
-        /// </returns>
-        public static string FindCondition(this ProjectElement projectElement)
+        public static class WellKnownPropertyNames
         {
-            if (projectElement == null)
-                throw new ArgumentNullException(nameof(projectElement));
-            
-            ProjectElement currentElement = projectElement;
-            while (currentElement != null)
-            {
-                if (!String.IsNullOrWhiteSpace(currentElement.Condition))
-                    return currentElement.Condition;
+            /// <summary>
+            ///     The "MSBuildExtensionsPath" property.
+            /// </summary>
+            public static readonly string MSBuildExtensionsPath = "MSBuildExtensionsPath";
 
-                currentElement = currentElement.Parent;
-            }
+            /// <summary>
+            ///     The "MSBuildSDKsPath" property.
+            /// </summary>
+            public static readonly string MSBuildSDKsPath = "MSBuildSDKsPath";
 
-            return String.Empty;
+            /// <summary>
+            ///     The "SolutionDir" property.
+            /// </summary>
+            public static readonly string SolutionDir = "SolutionDir";
+
+            /// <summary>
+            ///     The "_ResolveReferenceDependencies" property.
+            /// </summary>
+            public static readonly string ResolveReferenceDependencies = "_ResolveReferenceDependencies";
+
+            /// <summary>
+            ///     The "DesignTimeBuild" property.
+            /// </summary>
+            public static readonly string DesignTimeBuild = "DesignTimeBuild";
+
+            /// <summary>
+            ///     The "BuildProjectReferences" property.
+            /// </summary>
+            public static readonly string BuildProjectReferences = "BuildProjectReferences";
+
+            /// <summary>
+            ///     The "RoslynTargetsPath" property.
+            /// </summary>
+            public static readonly string RoslynTargetsPath = "RoslynTargetsPath";
         }
-
-        /// <summary>
-    ///     The names of well-known MSBuild properties.
-    /// </summary>
-    public static class WellKnownPropertyNames
-    {
-        /// <summary>
-        ///     The "MSBuildExtensionsPath" property.
-        /// </summary>
-        public static readonly string MSBuildExtensionsPath = "MSBuildExtensionsPath";
-
-        /// <summary>
-        ///     The "MSBuildSDKsPath" property.
-        /// </summary>
-        public static readonly string MSBuildSDKsPath = "MSBuildSDKsPath";
-
-        /// <summary>
-        ///     The "SolutionDir" property.
-        /// </summary>
-        public static readonly string SolutionDir = "SolutionDir";
-
-        /// <summary>
-        ///     The "_ResolveReferenceDependencies" property.
-        /// </summary>
-        public static readonly string ResolveReferenceDependencies = "_ResolveReferenceDependencies";
-
-        /// <summary>
-        ///     The "DesignTimeBuild" property.
-        /// </summary>
-        public static readonly string DesignTimeBuild = "DesignTimeBuild";
-
-        /// <summary>
-        ///     The "BuildProjectReferences" property.
-        /// </summary>
-        public static readonly string BuildProjectReferences = "BuildProjectReferences";
-
-        /// <summary>
-        ///     The "RoslynTargetsPath" property.
-        /// </summary>
-        public static readonly string RoslynTargetsPath = "RoslynTargetsPath";
-    }
     }
 }
