+++ src/LanguageServer.Common/Utilities/NuGetHelper.cs
