+++ src/LanguageServer.Engine/Utilities/ModelConversions.cs
@@ -1,10 +1,7 @@
-using System;
-using Microsoft.Build.Construction;
-
-using MLXML = Microsoft.Language.Xml;
-
 namespace MSBuildProjectTools.LanguageServer.Utilities
 {
+    using SemanticModel;
+
     /// <summary>
     ///     Extension methods for converting models between native and third-party representations.
     /// </summary>
@@ -54,26 +51,6 @@ namespace MSBuildProjectTools.LanguageServer.Utilities
         }
 
         /// <summary>
-        ///     Convert the MSBuild <see cref="ElementLocation"/> to its native equivalent.
-        /// </summary>
-        /// <param name="location">
-        ///     The <see cref="ElementLocation"/> to convert.
-        /// </param>
-        /// <returns>
-        ///     The equivalent <see cref="Position"/>.
-        /// </returns>
-        public static Position ToNative(this ElementLocation location)
-        {
-            if (location == null)
-                return null;
-
-            if (location.Line == 0)
-                return Position.Invalid;
-
-            return new Position(location.Line, location.Column);
-        }
-
-        /// <summary>
         ///     Convert the <see cref="Range"/> to its Language Server Protocol equivalent.
         /// </summary>
         /// <param name="range">
@@ -91,31 +68,6 @@ namespace MSBuildProjectTools.LanguageServer.Utilities
                 range.Start.ToLsp(),
                 range.End.ToLsp()
             );
-        }
-
-        /// <summary>
-        ///     Convert the <see cref="MLXML.TextSpan"/> to its native equivalent.
-        /// </summary>
-        /// <param name="span">
-        ///     The <see cref="MLXML.TextSpan"/> to convert.
-        /// </param>
-        /// <param name="textPositions">
-        ///     The textual position lookup used to map absolute positions to lines and columns.
-        /// </param>
-        /// <returns>
-        ///     The equivalent <see cref="Range"/>.
-        /// </returns>
-        public static Range ToNative(this MLXML.TextSpan span, TextPositions textPositions)
-        {
-            if (textPositions == null)
-                throw new ArgumentNullException(nameof(textPositions));
-
-            Position startPosition = textPositions.GetPosition(span.Start);
-            Position endPosition = textPositions.GetPosition(span.End);
-            if (endPosition.ColumnNumber == 0)
-                throw new InvalidOperationException("Should not happen anymore");
-
-            return new Range(startPosition, endPosition);
         }
 
         /// <summary>
