+++ src/LanguageServer.SemanticModel.MSBuild/MSBuildItemGroup.cs
