+++ src/LanguageServer.SemanticModel.MSBuild/MSBuildLocator.cs
@@ -4,7 +4,6 @@ using Microsoft.Language.Xml;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Xml.Linq;
 
 namespace MSBuildProjectTools.LanguageServer.SemanticModel
 {
