+++ src/LanguageServer.SemanticModel.MSBuild/MSBuildObject.cs
