+++ src/LanguageServer.SemanticModel.MSBuild/MSBuildObjectKind.cs
