+++ src/LanguageServer.SemanticModel.MSBuild/MSBuildProperty.cs
