+++ src/LanguageServer.SemanticModel.MSBuild/MSBuildUnusedItemGroup.cs
