+++ src/LanguageServer.SemanticModel.MSBuild/MSBuildUnusedProperty.cs
