+++ src/LanguageServer.SemanticModel.Xml/XSAttribute.cs
