+++ src/LanguageServer.SemanticModel.Xml/XSElement.cs
