+++ src/LanguageServer.SemanticModel.Xml/XSElementText.cs
