+++ src/LanguageServer.SemanticModel.Xml/XSElementWithContent.cs
