+++ src/LanguageServer.SemanticModel.Xml/XSEmptyElement.cs
