+++ src/LanguageServer.SemanticModel.Xml/XSInvalidAttribute.cs
