+++ src/LanguageServer.SemanticModel.Xml/XSInvalidElement.cs
