+++ src/LanguageServer.SemanticModel.Xml/XSNode.cs
