+++ src/LanguageServer.SemanticModel.Xml/XSParser.cs
