+++ src/LanguageServer.SemanticModel.Xml/XSWhitespace.cs
