+++ src/LanguageServer.SemanticModel.Xml/XmlPosition.cs
