+++ src/LanguageServer.SemanticModel.Xml/XmlPositionFlags.cs
