+++ src/LanguageServer.SemanticModel/LanguageServer.SemanticModel.csproj
@@ -1,0 +1,9 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <AssemblyName>MSBuildProjectTools.LanguageServer.SemanticModel</AssemblyName>
+    <RootNamespace>MSBuildProjectTools.LanguageServer.SemanticModel</RootNamespace>
+  </PropertyGroup>
+
+</Project>
