+++ test/LanguageServer.Engine.Tests/LanguageServer.Engine.Tests.csproj
@@ -33,7 +33,10 @@
     </ItemGroup>
     
     <ItemGroup>
-        <ProjectReference Include="../../src/LanguageServer.Engine/LanguageServer.Engine.csproj" />
+      <ProjectReference Include="../../src/LanguageServer.Common/LanguageServer.Common.csproj" />
+      <ProjectReference Include="../../src/LanguageServer.Engine/LanguageServer.Engine.csproj" />
+      <ProjectReference Include="../../src/LanguageServer.SemanticModel.MSBuild/LanguageServer.SemanticModel.MSBuild.csproj" />
+      <ProjectReference Include="../../src/LanguageServer.SemanticModel.Xml/LanguageServer.SemanticModel.Xml.csproj" />
     </ItemGroup>
 
     <ItemGroup>
