+++ test/LanguageServer.Engine.Tests/RangeTests.cs
@@ -5,6 +5,8 @@ using Xunit;
 
 namespace MSBuildProjectTools.LanguageServer.Tests
 {
+    using SemanticModel;
+
     /// <summary>
     ///     Tests for <see cref="Range"/>.
     /// </summary>
